refactor: move MenuOptions to /SideMenu folder
